front/fullstack/training-1/back/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
greenlet>=3.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9